+++ example/example/lib/story_book_widget.dart
@@ -23,8 +23,10 @@ class StoryBookWidget extends StatelessWidget {
               width: 300,
               decoration: BoxDecoration(color: ZwapColors.neutral200, borderRadius: BorderRadius.circular(20)),
               padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 20),
-              child: Column(
-                children: stories.map((e) => _getStorySelector(e, () => context.read<StoryProvider>().currentStory = e.story)).toList(),
+              child: SingleChildScrollView(
+                child: Column(
+                  children: stories.map((e) => _getStorySelector(e, () => context.read<StoryProvider>().currentStory = e.story)).toList(),
+                ),
               ),
             ),
             Expanded(child: Builder(builder: stories.firstWhere((s) => s.story == _currentStory).builder)),